Support local bot api and webhook

+++ core/os_util.go
@@ -33,18 +33,36 @@ func purgeOutdatedStorageData() {
 		}
 	}
 
-	webappDataDirEntries, _ := os.ReadDir(Config.WebappDataDir)
-	for _, f := range webappDataDirEntries {
-		if !f.IsDir() {
-			continue
+	if Config.WebappDataDir != "" {
+		webappDataDirEntries, _ := os.ReadDir(Config.WebappDataDir)
+		for _, f := range webappDataDirEntries {
+			if !f.IsDir() {
+				continue
+			}
+			fInfo, _ := f.Info()
+			fMTime := fInfo.ModTime().Unix()
+			fPath := filepath.Join(Config.WebappDataDir, f.Name())
+			// 2 Days
+			if fMTime < (time.Now().Unix() - 172800) {
+				os.RemoveAll(fPath)
+				log.Infoln("Purged outdated webapp data dir:", fPath)
+			}
 		}
-		fInfo, _ := f.Info()
-		fMTime := fInfo.ModTime().Unix()
-		fPath := filepath.Join(Config.WebappDataDir, f.Name())
-		// 2 Days
-		if fMTime < (time.Now().Unix() - 172800) {
-			os.RemoveAll(fPath)
-			log.Infoln("Purged outdated webapp data dir:", fPath)
-		}
+	}
+
+	if Config.LocalBotApiDir != "" {
+		filepath.Walk(Config.LocalBotApiDir, func(path string, info os.FileInfo, err error) error {
+			if info.IsDir() {
+				return nil
+			}
+			// fInfo, _ := f.Info()
+			fMTime := info.ModTime().Unix()
+			// 2 Days
+			if fMTime < (time.Now().Unix() - 172800) {
+				os.RemoveAll(path)
+				log.Infoln("Purged outdated LocalBotApiDir:", path)
+			}
+			return nil
+		})
 	}
 }
